Refaktor. Trekker ut fetchFn
