+++ src/frontend/utils/fetch.ts
@@ -1,14 +1,47 @@
+import { v4 as uuidv4 } from 'uuid';
+
 import { RessursFeilet, RessursStatus, RessursSuksess } from '../typer/ressurs';
 
-export const håndterSuksess = <ResponseData>(
-    res: Response
-): Promise<RessursSuksess<ResponseData>> =>
+export const fetchFn = <ResponseData, RequestData>(
+    url: string,
+    method: 'GET' | 'POST' | 'PUT' | 'PATCH' | 'DELETE' | 'UPDATE' = 'GET',
+    data: RequestData
+): Promise<RessursSuksess<ResponseData> | RessursFeilet> => {
+    const requestId = uuidv4().replaceAll('-', '');
+
+    return fetch(url, {
+        body: JSON.stringify(data),
+        method: method,
+        headers: {
+            'Content-Type': 'application/json',
+            'x-request-id': requestId,
+        },
+    })
+        .then((res): Promise<RessursSuksess<ResponseData> | RessursFeilet> => {
+            if (res.ok) {
+                return håndterSuksess<ResponseData>(res);
+            } else {
+                return håndterFeil(res, res.headers);
+            }
+        })
+        .catch((error) => {
+            // TODO: sjekk for 401
+
+            return {
+                status: RessursStatus.FEILET,
+                frontendFeilmelding: feilmeldingMedCallId(error.detail, error.headers),
+                melding: error.detail,
+            };
+        });
+};
+
+const håndterSuksess = <ResponseData>(res: Response): Promise<RessursSuksess<ResponseData>> =>
     res.json().then((data) => ({
         data: data as ResponseData,
         status: RessursStatus.SUKSESS,
     }));
 
-export const håndterFeil = (res: Response, headers: Headers): Promise<RessursFeilet> =>
+const håndterFeil = (res: Response, headers: Headers): Promise<RessursFeilet> =>
     res.json().then((res) => {
         return {
             status: RessursStatus.FEILET,
@@ -17,7 +50,7 @@ export const håndterFeil = (res: Response, headers: Headers): Promise<RessursFe
         };
     });
 
-export const feilmeldingMedCallId = (feilmelding: string, headers?: Headers): string => {
+const feilmeldingMedCallId = (feilmelding: string, headers?: Headers): string => {
     const callId = headers?.get('Nav-Call-id');
     return `${feilmelding}. Feilkode: ${callId}`;
 };
